Validate JSON files as part of linting process
Note that the JSON Lint CLI only accepts a single file and does not
support globbing [1].  This required writing a custom NPM script to
perform the linting instead of invoking the jsonlint command directly.
This script can be replaced if multiple file support is added to JSON
Lint in the future.

[1] https://github.com/zaach/jsonlint/issues/77