Fix tests broken after the upgrade to ESLint 4
The "should not report a violation when only spaces are used" test is
broken in ESLint 4 due to stricter checking on expected indentation.
The test now fails because the source under test should start in column
0.  Given that the test is effectively ill-posed and the requirements
it's testing are covered by several other tests in this fixture, it has
been removed.

The "should report a violation when both spaces and tabs are used" test,
while not broken, was updated to ensure it didn't fail for a different
violation in a future version of ESLint.  That is, it may yet be
interpreted to fail because the source under test did not start in
column 0 depending on the order in which rules are evaluated.
Therefore, a source input was chosen that is otherwise valid according
to the other rules in this configuration.